Split AI chat runtime reset send tests
